refactor(#17): move suggested tags list component to shared dir
