+++ src/app/App/NoteEditor/index.tsx
@@ -11,7 +11,7 @@ import { ITag } from '../../../core/Registry/Tags/Tags';
 import { $tags, setActiveTag, tagAttachmentsChanged, tagsChanged } from '../../../core/state/tags';
 import { Icon } from '../../components/Icon/Icon.bundle/common';
 
-import { SuggestedTagsList } from '../MainScreen/NotesOverview/TagEditor/SuggestedTagsList';
+import { SuggestedTagsList } from '../../components/SuggestedTagsList';
 import { FileUploader } from '../MonakoEditor/features/useDropFiles';
 import { MonacoEditor } from '../MonakoEditor/MonacoEditor';
 import { NoteScreen } from '../NoteScreen';
